#1 refactored server answer and tests
