+++ assistant_dists/dream/docker-compose.override.yml
@@ -657,10 +657,10 @@ services:
         MODEL1_ARG: http://files.deeppavlov.ai/alexaprize_data/parlai_grounding_knowledge/topical_chat_blender90_1_sent_48_epochs.tar.gz
         MODEL2_ARG: http://files.deeppavlov.ai/alexaprize_data/parlai_grounding_knowledge/topical_chat_blender90_3_sent_62_epochs.tar.gz
       context: ./services/knowledge_grounding/
-    command: flask run -h 0.0.0.0 -p 8083
+    command: uvicorn server:app --host 0.0.0.0 --port 8083
     environment:
       - CUDA_VISIBLE_DEVICES=0
-      - FLASK_APP=server
+      - FASTAPI_APP=server
     deploy:
       resources:
         limits:
