+++ services/knowledge_grounding/requirements.txt
@@ -1,10 +1,8 @@
-flask==2.1.0
 itsdangerous==2.0.1
-gunicorn==19.9.0
 requests==2.22.0
-jinja2<=3.0.3
-sentry-sdk[flask]==0.14.1
-jinja2<=3.0.3
-Werkzeug<=2.0.3
+sentry-sdk[fastapi]
 markupsafe==2.0.1
 nltk==3.2.5
+fastapi[all]
+uvicorn[standard]
+pytest
