+++ services/knowledge_grounding/server.py
@@ -1,33 +1,40 @@
 import random
 import time
+from typing import List, Optional
 
 import sentry_sdk
-from flask import Flask, request, jsonify
+from fastapi import Body, FastAPI
+from pydantic import BaseModel
 
 from config import *
 
-app = Flask(__name__)
+app = FastAPI()
 
 
-@app.route("/respond", methods=["POST"])
-def respond():
-    batch = request.json["batch"]
-    responses = [""]
+class Answer(BaseModel):
+    text: str
+
+
+@app.post('/respond', response_model=Optional[List[Answer]])
+def respond(payload: dict = Body(...)):
+    batch = payload['batch']
+    responses = None
     random.seed(42)
     st_time = time.time()
     if batch:
-        user_inputs = {"history": batch[0]["history"].split("\n") if batch[0]["history"] else [""], "inputs": []}
-        for sample in batch:
-            user_inputs["inputs"].append(
-                {
-                    "checked_sentence": sample["checked_sentence"],
-                    "knowledge": sample["knowledge"],
-                    "text": sample["text"],
-                }
-            )
+        user_inputs = {
+            "history": batch[0]["history"].split("\n") if batch[0]["history"] else [""],
+            "inputs": [{
+                "checked_sentence": sample["checked_sentence"],
+                "knowledge": sample["knowledge"],
+                "text": sample["text"],
+            }
+                for sample in batch
+            ]
+        }
         try:
             raw_responses = kg_script.run(user_inputs)
-            responses = [r["text"] for r in raw_responses]
+            responses = [Answer(text=r["text"]) for r in raw_responses]
             logger.info(f"Current sample responses: {responses}")
         except Exception as e:
             sentry_sdk.capture_exception(e)
@@ -36,4 +43,4 @@ def respond():
         logger.info("Received empty batch, exiting with empty responses")
     total_time = time.time() - st_time
     logger.info(f"knowledge grounding one batch exec time: {total_time:.3f}s")
-    return jsonify(responses)
+    return responses
