+++ services/knowledge_grounding/tests/conftest.py
@@ -8,4 +8,3 @@ def pytest_addoption(parser):
 @pytest.fixture
 def url(request) -> str:
     return request.config.getoption('--url')
-
